Django-backend/requirements.txt
asgiref==3.6.0
colorama==0.4.6
Django==4.2.1
django-cors-headers==3.14.0
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
psycopg2-binary==2.9.6
PyJWT==2.6.0
pytest==7.3.1
pytest-django==4.5.2
sqlparse==0.4.4
tzdata==2023.3
djangorestframework==3.14.0
gunicorn==21.2.0
drf-yasg==1.21.7
django-heroku==0.3.1
djangorestframework-simplejwt==5.3.0